refactor(web): rename version-or-new-file-dialog.logic to file-extension
The helpers (extensionMatches, getExtension) are generic file-extension
utilities; the dialog-suffixed name misled about ownership since
use-upload-version.ts also imports them. Rename the module and update
both call sites.
